requirements.txt
django
djangorestframework
python-decouple==3.8
huggingface_hub
djangorestframework-simplejwt
